cleaned up helpers
