+++ src/utils/helper.ts
@@ -1,7 +1,40 @@
+/*---------------------------------------------------------------------------------------------
+ *  Copyright (c) Microsoft Corporation. All rights reserved.
+ *  Licensed under the MIT License. See License.txt in the project root for license information.
+ *--------------------------------------------------------------------------------------------*/
+
 export function changeCamelToDash(str: string) {
     str = str.replace(/[A-Z]/g, letter => `-${letter.toLowerCase()}`);
     if(str.startsWith('-')) {
         str = str.substring(1, str.length);
     }
     return str;
+}
+
+export function ToSnakeCase(v: string)
+{
+    return v.replace(/([a-z](?=[A-Z]))/g, '$1 ').split(' ').join('_').toLowerCase();
+}
+
+export function Capitalize(v: string) {
+    return v.charAt(0).toUpperCase() + v.slice(1);
+}
+
+export function Uncapitalize(v: string) {
+    return v.charAt(0).toLowerCase() + v.slice(1);
+}
+
+export function ToCamelCase(v: string)
+{
+    v = v.toLowerCase().replace(/[^A-Za-z0-9]/g, ' ').split(' ')
+    .reduce((result, word) => result + Capitalize(word.toLowerCase()));
+    return v.charAt(0).toLowerCase() + v.slice(1);
+}
+
+export function EscapeString(original: string): string
+{
+    if (original == undefined) return "undefined";
+    original = original.split('\n').join(" ");
+    original = original.split('\'').join("\\\'");
+    return original;
 }
